pyproject.toml
[tool.poetry]
name = "task"
version = "0.1.0"
description = ""
authors = ["Serhii Hlushak <glushaksergey90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.13.1"
bs4 = "^0.0.2"
phonenumbers = "^8.13.48"
tldextract = "^5.1.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
